Relocate refresh controllers to AdminController
